fix(agent): split replay transport files under revive line limit
process_transport_replay.go grew past the 800-line revive cap after the
synthetic optional-probe path; move recv and synthetic helpers out.
